Move deep merge logic to separate file
